This the Analysis of the ERP
The purpose of this work is to create our own ERP, for a Secondary Education Tutoring Center, this is done in the context of the course "Enterprise Resource Management Software Systems (Erp)" for the Master's degree in Computer Science at the University of Piraeus. At first, we will mention some information about ERPs and in the following chapters we will analyze the requirements of our ERP and the detailed design of the System. At the end of this paper we hope that the reader and ourselves will have a better understanding of the concept of ERPs and the difficulties we encounter in their construction.